Move test helpers to a sub dir
