+++ tests/test_helpers/flash_message_helper.go
@@ -1,4 +1,4 @@
-package tests
+package test_helpers
 
 import (
 	"net/http"
